requirements.txt
# AI Resume Analyzer - Python Dependencies
# Install with: pip install -r requirements.txt

flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
PyMuPDF==1.23.8
python-docx==1.1.0
werkzeug==3.0.1
torch>=2.0.0
